🚚[Rename] CARD_CQ_WIDTH 폴더 위치 변경
