+++ src/storybook/constants.ts
@@ -1,0 +1,11 @@
+/**
+ * 스토리북 공용 CQ 너비 상수
+ */
+
+export const CARD_CQ_WIDTHS = [
+  { label: "180px", desc: "@card-xs", width: 180 },
+  { label: "220px", desc: "@card-sm", width: 220 },
+  { label: "280px", desc: "@card-md", width: 280 },
+  { label: "320px", desc: "@card-lg", width: 320 },
+  { label: "380px", desc: "> @card-lg", width: 380 },
+] as const;
